requirements.txt
Django==2.2.20
djangorestframework==3.12.4
pkg-resources==0.0.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
